requirements.txt
Flask==3.0.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.1
APScheduler==3.10.4
selenium==4.15.2
webdriver-manager==4.0.1
Pillow==10.1.0
lxml==4.9.3
twilio==9.0.0
requests==2.31.0
APScheduler==3.10.4
google-auth==2.22.0
Flask-Mail
